Select.storiesの修正
